+++ buckaroo/dataflow/dataflow.py
@@ -3,10 +3,12 @@ import sys
 import warnings
 import pandas as pd
 from traitlets import Unicode, Any, observe, HasTraits, Dict
-from .serialization_utils import pd_to_obj    
+from ..serialization_utils import pd_to_obj    
+from buckaroo.pluggable_analysis_framework.utils import (filter_analysis)
 from buckaroo.pluggable_analysis_framework.pluggable_analysis_framework import (ColAnalysis)
 from buckaroo.pluggable_analysis_framework.analysis_management import DfStats
-from .customizations.all_transforms import configure_buckaroo, DefaultCommandKlsList
+
+from ..customizations.all_transforms import configure_buckaroo, DefaultCommandKlsList
 
 SENTINEL_DF_1 = pd.DataFrame({'foo'  :[10, 20], 'bar' : ["asdf", "iii"]})
 SENTINEL_DF_2 = pd.DataFrame({'col1' :[55, 55], 'col2': ["pppp", "333"]})
@@ -14,13 +16,23 @@ SENTINEL_DF_3 = pd.DataFrame({'pp'   :[99, 33], 'ee':   [     6,     9]})
 SENTINEL_DF_4 = pd.DataFrame({'vvv'  :[12, 49], 'oo':   [ 'ccc', 'www']})
 
 
-
-
 def merge_ops(existing_ops, cleaning_ops):
     """ strip cleaning_ops from existing_ops, reinsert cleaning_ops at the beginning """
     a = existing_ops.copy()
     a.extend(cleaning_ops)
     return a
+
+def merge_sds(*sds):
+    """merge sds with later args taking precedence
+
+    sub-merging of "overide_config"??
+    """
+    base_sd = {}
+    for sd in sds:
+        for column in sd.keys():
+            base_sd[column] = merge_column(base_sd.get(column, {}), sd[column])
+    return base_sd
+
 
 def merge_column(base, new):
     """
@@ -36,17 +48,6 @@ def merge_column(base, new):
     if len(base_override) > 0:
         ret['column_config_override'] = base_override
     return ret
-
-def merge_sds(*sds):
-    """merge sds with later args taking precedence
-
-    sub-merging of "overide_config"??
-    """
-    base_sd = {}
-    for sd in sds:
-        for column in sd.keys():
-            base_sd[column] = merge_column(base_sd.get(column, {}), sd[column])
-    return base_sd
 
 SENTINEL_COLUMN_CONFIG_1 = "ASDF"
 SENTINEL_COLUMN_CONFIG_2 = "FOO-BAR"
@@ -118,11 +119,12 @@ class DataFlow(HasTraits):
         super().__init__()
         self.summary_sd = {}
         self.existing_operations = []
-
         try:
             self.raw_df = raw_df
         except Exception:
             six.reraise(self.exception[0], self.exception[1], self.exception[2])
+
+    #autocleaning_klass = SentinelAutocleaning
 
     raw_df = Any('')
     sample_method = Unicode('default')
@@ -201,6 +203,7 @@ class DataFlow(HasTraits):
         cleaned_df = self._run_df_interpreter(self.sampled_df, merged_operations)
         generated_code = self._run_code_generator(merged_operations)
         self.cleaned = [cleaned_df, cleaning_sd, generated_code, merged_operations]
+
 
     @property
     def cleaned_df(self):
@@ -335,13 +338,6 @@ class StylingAnalysis(ColAnalysis):
             'column_config': ret_col_config}
 
 
-def filter_analysis(klasses, attr):
-    ret_klses = {}
-    for k in klasses:
-        attr_val = getattr(k, attr, None)
-        if attr_val is not None:
-            ret_klses[attr_val] = k
-    return ret_klses
             
 class PostProcessingException(Exception):
     pass
@@ -379,7 +375,6 @@ class CustomizableDataflow(DataFlow):
     This allows targetd extension and customization of DataFlow
     """
     analysis_klasses = [StylingAnalysis]
-    command_klasses = DefaultCommandKlsList
     commandConfig = Dict({}).tag(sync=True)
     DFStatsClass = DfStats
     sampling_klass = Sampling
@@ -400,7 +395,6 @@ class CustomizableDataflow(DataFlow):
         self.setup_options_from_analysis()
         super().__init__(self.sampling_klass.pre_stats_sample(df))
 
-        self._setup_from_command_kls_list()
         self.populate_df_meta()
         #self.raw_df = df
         warnings.filterwarnings('default')
@@ -449,34 +443,6 @@ class CustomizableDataflow(DataFlow):
     #empty needs to always be present, it enables startup
     df_data_dict = Any({'empty':[]}).tag(sync=True)
 
-
-    ### start code interpreter block
-    def _setup_from_command_kls_list(self):
-        #used to initially setup the interpreter, and when a command
-        #is added interactively
-        c_klasses = self.command_klasses
-        c_defaults, c_patterns, df_interpreter, gencode_interpreter = configure_buckaroo(c_klasses)
-        self.df_interpreter, self.gencode_interpreter = df_interpreter, gencode_interpreter
-        self.commandConfig = dict(argspecs=c_patterns, defaultArgs=c_defaults)
-
-    def add_command(self, incomingCommandKls):
-        without_incoming = [x for x in self.command_classes if not x.__name__ == incomingCommandKls.__name__]
-        without_incoming.append(incomingCommandKls)
-        self.command_klasses = without_incoming
-        self.setup_from_command_kls_list()
-
-    def _run_df_interpreter(self, df, operations):
-        full_ops = [{'symbol': 'begin'}]
-        full_ops.extend(operations)
-        if len(full_ops) == 1:
-            return df
-        return self.buckaroo_transform(full_ops , df)
-
-    def run_code_generator(self, operations):
-        if len(operations) == 0:
-            return 'no operations'
-        return self.gencode_interpreter(operations)
-    ### end code interpeter block
 
     def _compute_processed_result(self, cleaned_df, post_processing_method):
         if post_processing_method == '':
